refactor(gateway): drop dead envelope fields and stt scaffolding
Remove envelope fields nothing reads (Attachment.mime/caption,
ReplyContext.message_id, InboundMessage.timestamp, account_id), the
unreachable SttProvider.available()/unavailable_message() pair and the
Transcriber protocol, fold EchoTurnHandler into the echo adapter, and fix
stale docstrings. Capability flags and send_media stay for future adapters.
